chore(config): (0)
move theme helpers to config
