+++ src/config/index.ts
@@ -1,0 +1,4 @@
+import * as constants from './constants';
+export { default as config } from './config';
+
+export { constants };
