+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './utils';
-export { default as config } from './config';
+export { config } from './config';
 export { Container } from './components/Container';
